# docs/swagger.json
{
	"components": {
...
		"schemas": {
			"Input": {
				"properties": {
					"name": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"email",
							"password",
							"text",
							"mobile",
							"date",
							"time",
							"datetime",
							"radio",
							"select",
							"file",
							"button",
							"radiogroup",
							"autocomplete",
							"emailOrMobile",
							"checkbox",
							"textarea",
							"number",
							"switch",
							"slider",
							"rating",
							"color",
							"submit",
							"reset",
							"buttonGroup"
						]
					},
					"helperText": {
						"type": "string"
					},
					"Variant": {
						"type": "string",
						"enum": [
							"outline",
							"standard",
							"contained",
							"text"
						]
					},
					"sx": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					"fullWidth": {
						"type": "boolean"
					},
					"validation": {},
					"onClick": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"label",
					"type"
				],
				"type": "object",
				"additionalProperties": {}
			},
			"StandardForm": {
				"properties": {
					"title": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"inputs": {
						"items": {
							"$ref": "#/components/schemas/Input"
						},
						"type": "array"
					},
					"btnTxt": {
						"type": "string"
					},
					"noSubmit": {
						"type": "boolean"
					},
					"onSubmitSuccess": {
						"type": "string"
					},
					"onSubmitFailure": {
						"type": "string"
					},
					"beforeSubmit": {
						"type": "string"
					},
					"clearForm": {
						"type": "string"
					}
				},
				"required": [
					"inputs"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICreate": {
				"properties": {
					"form": {
						"$ref": "#/components/schemas/StandardForm"
					},
					"header": {
						"type": "string"
					},
					"sx": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					"activationDate": {
						"type": "string"
					},
					"expiryDate": {
						"type": "string"
					},
					"social": {
						"properties": {
							"telegram": {
								"type": "boolean"
							},
							"whatsapp": {
								"type": "boolean"
							}
						},
						"required": [
							"telegram",
							"whatsapp"
						],
						"type": "object"
					},
					"webHooks": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"collectionGroup": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					}
				},
				"required": [
					"form",
					"header",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUpdate": {
				"properties": {
					"form": {
						"$ref": "#/components/schemas/StandardForm"
					},
					"header": {
						"type": "string"
					},
					"sx": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					"activationDate": {
						"type": "string"
					},
					"expiryDate": {
						"type": "string"
					},
					"social": {
						"properties": {
							"telegram": {
								"type": "boolean"
							},
							"whatsapp": {
								"type": "boolean"
							}
						},
						"required": [
							"telegram",
							"whatsapp"
						],
						"type": "object"
					},
					"webHooks": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"collectionGroup": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"formId": {
						"type": "string"
					}
				},
				"required": [
					"form",
					"header",
					"userId",
					"formId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IDelete": {
				"properties": {
					"formId": {
						"type": "string"
					}
				},
				"required": [
					"formId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "questionia",
		"version": "1.0.0",
		"description": "a basic express app writen in typescript",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Abdulraheem A omeiza"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/form/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"success": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"409": {
						"description": "Failed to create form"
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Forms"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreate"
							}
						}
					}
				}
			}
		},
		"/form/{id}": {
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"success": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"409": {
						"description": "Failed to update form"
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Forms"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUpdate"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"success": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"409": {
						"description": "Failed to delete form"
					}
				},
				"tags": [
					"Forms"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IDelete"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"success": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"409": {
						"description": "Failed to create form"
					}
				},
				"tags": [
					"Forms"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/form": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"success": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"409": {
						"description": "Failed to fetch forms"
					}
				},
				"tags": [
					"Forms"
				],
				"security": [],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}